Move StopSearchBar into components folder
