+++ ui/src/components/stop-registry/search/components/StopSearchBar/SearchForDropdown.tsx
@@ -4,9 +4,9 @@ import React, { FC, ReactNode } from 'react';
 import { useController } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
-import { ListboxButton, dropdownTransition } from '../../../../uiComponents';
-import { InputLabel } from '../../../forms/common';
-import { SearchFor, StopSearchFilters } from '../types';
+import { ListboxButton, dropdownTransition } from '../../../../../uiComponents';
+import { InputLabel } from '../../../../forms/common';
+import { SearchFor, StopSearchFilters } from '../../types';
 
 const disabled: ReadonlyArray<SearchFor> = [
   SearchFor.StopAreas,
