+++ ui/src/components/stop-registry/search/components/StopSearchBar/index.ts
