+++ ui/src/components/stop-registry/search/components/index.ts
@@ -1,3 +1,4 @@
 export * from './LoadingStopsErrorRow';
 export * from './LoadingStopsRow';
 export * from './StopGroupSelector';
+export * from './StopSearchBar';
