+++ ui/src/components/stop-registry/search/index.ts
@@ -1,4 +1,4 @@
-export * from './StopSearchBar';
+export { StopSearchBar } from './components';
 export * from './StopSearchResultsPage';
 export * from './StopTableRow/StopTableRow';
 export * from './types';
